+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # rcc change log
 
+## v17.21.0 (date: 25.3.2024) WORK IN PROGRESS
+
+- experimental feature to disable compression of hololib content
+- made cleanup much more strict on error detection
+- bug fix: task shell was missing `--space` option; added now
+
 ## v17.20.0 (date: 12.3.2024) WORK IN PROGRESS
 
 - uv experiment with limited scope and imperfect implementation